+++ database/migrations/2025_05_04_042220_create_invest_smart_transactions_table.php
@@ -11,18 +11,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('invest_smart_transactions', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('portfolio_id')->constrained('invest_smart_portfolios')->onDelete('cascade');
-            $table->enum('type', ['buy', 'sell']);
-            $table->string('symbol', 10);
-            $table->string('name');
-            $table->integer('quantity');
-            $table->decimal('price', 15, 2);
-            $table->decimal('total', 15, 2);
-            $table->timestamp('transaction_date');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('invest_smart_transactions')) {
+            Schema::create('invest_smart_transactions', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('portfolio_id')->constrained('invest_smart_portfolios')->onDelete('cascade');
+                $table->enum('type', ['buy', 'sell']);
+                $table->string('symbol', 10);
+                $table->string('name');
+                $table->integer('quantity');
+                $table->decimal('price', 15, 2);
+                $table->decimal('total', 15, 2);
+                $table->timestamp('transaction_date');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
